// frontend/src/lib/api/generated.ts
export type BlockResponseDtoContent = { [key: string]: unknown };

export type BlockResponseDtoProps = { [key: string]: unknown };

export interface BlockResponseDto {
  id: string;
  type: string;
  content: BlockResponseDtoContent;
  props?: BlockResponseDtoProps;
  order: number;
  pageId: string;
  createdAt: string;
  updatedAt: string;
}